Module oak_channel::basic_framed

Basic framed format for use before the channel is handed off to the application.

Depending on configuration, this protocol is used twice.

To load the application. Here the host is the sender, enclave is the recipient.

[Optional]: to return the attestation evidence. Here the host is the recipient, enclave is the sender.

The protocol is very simple:

sender sends the size of the payload, as u32

sender sends MAX_SIZE of data and waits for the recipient to acknowledge

recipient reads up to MAX_SIZE of data and acks by responding with the amount of data read

repeat (2) and (3) until all the data has been transmitted

Functions

receive_raw

send_raw