Module oak_channel::message

Implements the message layer as defined in /oak_channel/SPEC.md.

Structs

RequestMessage

Rust implementation of the Request Message structure defined in /oak_channel/SPEC.md

ResponseMessage

Rust implementation of the Response Message structure defined in /oak_channel/SPEC.md.

Constants

BODY_OFFSET

INVOCATION_ID_OFFSET

INVOCATION_ID_SIZE

LENGTH_OFFSET

LENGTH_SIZE

Traits

Message

Type Aliases

InvocationId

Length